Home

LiveDisplay.jl

LiveDisplay.liveinclude

LiveDisplay.@livedocs

LiveDisplay.liveinclude — Method.

liveinclude([module,] path; display=nothing)

Include path to module (default: Main) everytime it is updated.

Keyword Arguments

display: A callable that accepts one argument or a Vector of such callables. The object returned from include(path) is passed

to it/them.

source

LiveDisplay.@livedocs — Macro.

@livedocs THING [display=FUNCTION]

Display THING using FUNCTION (default to display) where THING can be:

Julia object with docstring.

Path to markdown file.

FUNCTION must be a callable that accepts an object returned by Base.Docs.doc (typically a Markdown.MD) or a Vector of such callables.

Examples

julia> @livedocs f

julia> using ElectronDisplay

julia> @livedocs "README.md" display=electrondisplay

source